Hot-fix FreeCommand ignoring messages from guild
After updating to JDA5, GuildMessageReceviedEvent was removed.
MessageReceivedEvent has to be used instead.

To "counter" this, one can use the isFromGuild method.

I use this method, but incorrect.
I should invert the if statemen; if the message is NOT from a guild, ignore.

This modification fixes the update that broke the help system.