feat: modularize sendRss() method and improve variable names
This also changes the functionality of how new RSS feeds get dealt with
for the first time. Before this commit, all items would get posted as
embeds on Discord, and as a result, that would bombard the target
channel with RSS posts once the routine executes.

This commit changes this behavior by assuming that all RSS posts have
been posted and it should only consider posting posts newer than the
registered date.