feat: infer roles from command parameters
After a discussion with the rest of the team, we decided that there are
plenty of flaws with having application roles stored in the
configuration file, primarily due to the fact that if somebody were to
want to change the roles displayed on the dropdown menu, they would have
to perform the tedious work of updating the configuration and then
restarting the server, but we can do better.

With this commit, all the application roles are inferred straight from
the arguments that are passed from the member executing the command. A
restart will not be needed in case somebody wants to change the
available roles or add a new one.