feat: add delimiter in dropdown menu option IDs
This way we are able to add more roles in the future without running
into any character limits. Getting the last character of the menu option
ID is not the most effective way to go about this.

Adding the delimiter in the menu option IDs allows us to treat the ID of
the role better and in a safer way.