refactor: handle `roleString` from early on
Get a reference of `roleString` as early as possible, and be more
explicit about which argument we want to get from the arguments list.

At any given point (think, a future update to some library like JDA),
the list we are getting all of our arguments from could end up looking
from this:

    [1192015871509315, "Role String"]

to:

    [1192015871509315, "Role String", 24]

Therefore, making `args.getLast()` entirely obsolete. Being more
explicit about the argument number we would like to get reduces the
chances of something like this happening.

Suggested-by: Zabuzard <zabuza.dev@gmail.com>
Signed-off-by: Chris Sdogkos <work@chris-sdogkos.com>