application-create: hardcode certain config values
Certain configuration values pertaining to the application form creating
and handling for custom roles do not need to be there and are deemed
more appropriate as hardcoded values inside the code itself.

Remove the configuration entries from the `config.json.template` as well
as references to such keys in the codebase and introduce them as static
constants in the appropriate class files.

Suggested-by: Zabuzard <zabuza.dev@gmail.com>
Signed-off-by: Chris Sdogkos <work@chris-sdogkos.com>