generateRoleOptions: remove unnecessary parameter
As it currently is, the helper method `generateRoleOptions` takes one
`SlashCommandData` parameter and is used once throughout the entire
project, specifically in the same class it is defined.

In that one case, `generateRoleOptions` would take in `getData()` as its
input, a method that is accessible through the entire class due to the
simple fact that `CreateRoleApplicationCommand` inherits methods from
`SlashCommandAdapter`.

Modify the `generateRoleOptions` helper method to take no input and
inside it, store a reference of `getData()` for the method to make use
of.

Suggested-by: Zabuzard <zabuza.dev@gmail.com>
Signed-off-by: Chris Sdogkos <work@chris-sdogkos.com>