refactor: remove unneeded word plural helper method
The `selectWordFromCount` helper method, while serving a seemingly
helpful purpose, is only used once and is also private to be used in the
same class only.

Besides the fact that we can technically move it somewhere more
appropriate and increase its visibility so other classes can utilize it,
remove the helepr method and resort to ternary operators for that one
time we need to determine the plurarity of the word "minute" in our
`CreateRoleApplicationCommand` class.

Suggested-by: Zabuzard <zabuza.dev@gmail.com>
Signed-off-by: Chris Sdogkos <work@chris-sdogkos.com>