refactor: simplify variable naming
Rename the variable name of the class-wide `RoleApplicationSystemConfig`
to `config` and the variable name of `RoleApplicationHandler` to
`handler`, more simple and recognizable names. The original ones are too
verbose and unnecessarily yield long horizontal lines of code where
unnecessary.

These verbose names do work but only in the situation where there are
multiple configurations or handlers used in the same class.

Suggested-by: Zabuzard <zabuza.dev@gmail.com>
Signed-off-by: Chris Sdogkos <work@chris-sdogkos.com>