refactor(application-form): better exception handling
There are currently many undesired conditions the code could run into
when it comes to the `RoleApplicationHandler` which we would not desire
and we would not be prepared for, particularly in situations where:

    - We would somehow find ourselves handling application form
      submissions in a non-guild environment.

    - The arguments size including the clicked component ID as well as
      the role name (the title of that component ID in essence) would
      have an unexpected size.

    - The staff-monitored application submissions text channel would not
      be found by the bot.

We do not have entire control over these cases, but we need to handle
them properly and set ourselves up for success in the rare case they
happen.

Make `sendApplicationResult` throw an `IllegalArgumentException` with
descriptive messages if anything goes wrong, and log appropriate
stacktraces if anything goes wrong in the logger for easy debugging.

I thought about printing straight to the logger and not using any
`IllegalArgumentException` in the code (or at least, using a different
or maybe custom `Exception` but that would be too much), but with this
method, we get to see custom messages along with exceptions for each
individual case - all that without unnecessarily overcomplicating the
code and sacrificing its readability to a considerable extent.

Suggested-by: Zabuzard <zabuza.dev@gmail.com>
Signed-off-by: Chris Sdogkos <work@chris-sdogkos.com>